separate POST booking api and get form data to MVC
