+++ static/js/controller/booking.js
@@ -6,7 +6,11 @@ const bookingURL = "/api/booking";
 
 document.addEventListener("DOMContentLoaded", async function () {
   const trashBtn = document.querySelector(".trash");
-  trashBtn.addEventListener("click", fetchDeleteBooking);
+  if (trashBtn) {
+    trashBtn.addEventListener("click", fetchDeleteBooking);
+  } else {
+    console.log("Trash button not found");
+  }
 
   if (window.location.pathname === "/booking") {
     const isLoggedIn = await checkUserState();
@@ -72,5 +76,31 @@ async function fetchDeleteBooking() {
     window.location.reload();
   } catch (error) {
     console.error("Error deleting attraction:", error);
+  }
+}
+
+export async function fetchPostBooking(bookingData) {
+  try {
+    const token = localStorage.getItem("userToken");
+    const response = await fetch(bookingURL, {
+      method: "POST",
+      headers: {
+        Authorization: `Bearer ${token}`,
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(bookingData),
+    });
+
+    if (!response.ok) {
+      window.location = "/";
+    }
+
+    const data = await response.json();
+    if (!data || !data.data || bookingData.attractionId != data.data.id) {
+      window.location = "/";
+    }
+    window.location.href = "/booking";
+  } catch (error) {
+    console.error("Error fetching post booking:", error);
   }
 }
